+++ src/latte/layers/dummy_data_layer.cc
@@ -55,7 +55,7 @@ void DummyDataLayer<Dtype>::LayerSetUp(const vector<Blob<Dtype> *> &bottom,
   this->Forward(bottom, top);
   // Invert the inverted refill_ values to refill the desired (non-constant)
   // Blobs in every usual forward pass.
-  for (int i = 0; i < refill_.size(); ++i) {
+  for (size_t i = 0; i < refill_.size(); ++i) {
     refill_[i] = !refill_[i];
   }
 }
@@ -63,7 +63,7 @@ void DummyDataLayer<Dtype>::LayerSetUp(const vector<Blob<Dtype> *> &bottom,
 template <typename Dtype>
 void DummyDataLayer<Dtype>::Forward_cpu(const vector<Blob<Dtype> *> &bottom,
                                         const vector<Blob<Dtype> *> &top) {
-  for (int i = 0; i < top.size(); ++i) {
+  for (size_t i = 0; i < top.size(); ++i) {
     const int filler_id = (fillers_.size() > 1) ? i : 0;
     if (refill_[filler_id]) {
       fillers_[filler_id]->Fill(top[i]);
