+++ src/latte/test/test_syncedmem.cc
@@ -15,7 +15,7 @@ class SyncedMemoryTest : public ::testing::Test {};
 TEST_F(SyncedMemoryTest, TestInitialization) {
   SyncedMemory mem(10);
   EXPECT_EQ(mem.head(), SyncedMemory::SyncedHead::UNINITIALIZED);
-  EXPECT_EQ(mem.size(), 10);
+  EXPECT_EQ(mem.size(), static_cast<size_t>(10));
   SyncedMemory *p_mem = new SyncedMemory(10 * sizeof(float));
   EXPECT_EQ(p_mem->size(), 10 * sizeof(float));
   delete p_mem;
@@ -54,14 +54,14 @@ TEST_F(SyncedMemoryTest, TestCPUWrite) {
   void *cpu_data = mem.mutable_cpu_data();
   EXPECT_EQ(mem.head(), SyncedMemory::SyncedHead::HEAD_AT_CPU);
   latte_memset(mem.size(), 1, cpu_data);
-  for (int i = 0; i < mem.size(); ++i) {
+  for (size_t i = 0; i < mem.size(); ++i) {
     EXPECT_EQ(reinterpret_cast<char*>(cpu_data)[i], 1);
   }
 
   cpu_data = mem.mutable_cpu_data();
   EXPECT_EQ(mem.head(), SyncedMemory::SyncedHead::HEAD_AT_CPU);
   latte_memset(mem.size(), 2, cpu_data);
-  for (int i = 0; i < mem.size(); ++i) {
+  for (size_t i = 0; i < mem.size(); ++i) {
     EXPECT_EQ(reinterpret_cast<char*>(cpu_data)[i], 2);
   }
 }
@@ -78,20 +78,20 @@ TEST_F(SyncedMemoryTest, TestGPURead) {
 
   char *recovered_value = new char[10];
   latte_gpu_memcpy(10, gpu_data, recovered_value);
-  for (int i = 0; i < mem.size(); ++i) {
+  for (size_t i = 0; i < mem.size(); ++i) {
     EXPECT_EQ(recovered_value[i], 1);
   }
 
   cpu_data = mem.mutable_cpu_data();
   EXPECT_EQ(mem.head(), SyncedMemory::SyncedHead::HEAD_AT_CPU);
   latte_memset(mem.size(), 2, cpu_data);
-  for (int i = 0; i < mem.size(); ++i) {
+  for (size_t i = 0; i < mem.size(); ++i) {
     EXPECT_EQ(reinterpret_cast<char*>(cpu_data)[i], 2);
   }
   gpu_data = mem.gpu_data();
   EXPECT_EQ(mem.head(), SyncedMemory::SyncedHead::SYNCED);
   latte_gpu_memcpy(10, gpu_data, recovered_value);
-  for (int i = 0; i < mem.size(); ++i) {
+  for (size_t i = 0; i < mem.size(); ++i) {
     EXPECT_EQ(recovered_value[i], 2);
   }
   delete[] recovered_value;
@@ -103,7 +103,7 @@ TEST_F(SyncedMemoryTest, TestGPUWrite) {
   EXPECT_EQ(mem.head(), SyncedMemory::SyncedHead::HEAD_AT_GPU);
   latte_gpu_memset(mem.size(), 1, gpu_data);
   const void *cpu_data = mem.cpu_data();
-  for (int i = 0; i < mem.size(); ++i) {
+  for (size_t i = 0; i < mem.size(); ++i) {
     EXPECT_EQ(reinterpret_cast<const char*>(cpu_data)[i], 1);
   }
 
@@ -111,7 +111,7 @@ TEST_F(SyncedMemoryTest, TestGPUWrite) {
   EXPECT_EQ(mem.head(), SyncedMemory::SyncedHead::HEAD_AT_GPU);
   latte_gpu_memset(mem.size(), 2, gpu_data);
   cpu_data = mem.cpu_data();
-  for (int i = 0; i < mem.size(); ++i) {
+  for (size_t i = 0; i < mem.size(); ++i) {
     EXPECT_EQ(reinterpret_cast<const char*>(cpu_data)[i], 2);
   }
   EXPECT_EQ(mem.head(), SyncedMemory::SyncedHead::SYNCED);
